Exercise 4.21*

+++ part4/controllers/blogs.js
@@ -40,8 +40,25 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  if (!request.token) {
+    response.status(401).end()
+  } else {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    
+    const blog = await Blog.findById(request.params.id)
+
+    if ( blog.user.toString() === user.id.toString() ) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(401).end()
+    }
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
